requirements.txt
numpy
torch
torchvision
tqdm
imageio
opencv-python
pandas
matplotlib
medpy
thop
pyyaml
